+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - #115 Fix non-latin names and MRNs comparison in a listing view
+- #114 Fix jsonapi returns None for sample's DateOfBirth field
 - #113 Added estimated birthdate field to Patient content type
 - #112 Use default TZ when calculating birthdate if `on_date` param is not set
 - #111 Allow/Disallow the introduction of future dates of birth